Move all the sized_number options into modules
