+++ src/datatype/composite/string/lpstring.rs
@@ -111,7 +111,7 @@ mod tests {
         let data = b"\x00\x07\x41\x42\xE2\x9D\x84\xE2\x98\xA2\xF0\x9D\x84\x9E\xF0\x9F\x98\x88\xc3\xb7".to_vec();
         let offset = Offset::Dynamic(Context::new(&data));
 
-        let size_type = H2Number::new(SizedDefinition::U16(Endian::Big), SizedDisplay::Decimal);
+        let size_type = H2Number::new(SizedDefinition::U16(Endian::Big), SizedDisplay::Decimal(Default::default()));
 
         let a = LPString::new(size_type, UTF8::new())?;
         assert_eq!("\"AB❄☢𝄞😈÷\"", a.to_display(offset)?);
@@ -124,7 +124,7 @@ mod tests {
         let data = b"\x00\x41".to_vec();
         let offset = Offset::Dynamic(Context::new(&data));
 
-        let size_type = H2Number::new(SizedDefinition::U8, SizedDisplay::Decimal);
+        let size_type = H2Number::new(SizedDefinition::U8, SizedDisplay::Decimal(Default::default()));
         let a = LPString::new(size_type, UTF8::new())?;
         assert_eq!("\"\"", a.to_display(offset)?);
 
@@ -136,7 +136,7 @@ mod tests {
         let data = b"".to_vec();
         let offset = Offset::Dynamic(Context::new(&data));
 
-        let size_type = H2Number::new(SizedDefinition::U8, SizedDisplay::Decimal);
+        let size_type = H2Number::new(SizedDefinition::U8, SizedDisplay::Decimal(Default::default()));
         let a = LPString::new(size_type, UTF8::new())?;
         assert!(a.to_display(offset).is_err());
 
@@ -148,7 +148,7 @@ mod tests {
         let data = b"\x00\x07PPPPPP\x41\x42\xE2\x9D\x84\xE2\x98\xA2\xF0\x9D\x84\x9E\xF0\x9F\x98\x88\xc3\xb7".to_vec();
         let offset = Offset::Dynamic(Context::new(&data));
 
-        let size_type = H2Number::new_aligned(Alignment::Loose(8), SizedDefinition::U16(Endian::Big), SizedDisplay::Decimal);
+        let size_type = H2Number::new_aligned(Alignment::Loose(8), SizedDefinition::U16(Endian::Big), SizedDisplay::Decimal(Default::default()));
 
         let a = LPString::new(size_type, UTF8::new())?;
         assert_eq!("\"AB❄☢𝄞😈÷\"", a.to_display(offset)?);
@@ -162,7 +162,7 @@ mod tests {
         let data = b"\x07\x41\x42\xE2\x9D\x84\xE2\x98\xA2\xF0\x9D\x84\x9E\xF0\x9F\x98\x88\xc3\xb7".to_vec();
         let offset = Offset::Dynamic(Context::new(&data));
 
-        let size_type = H2Number::new(SizedDefinition::U8, SizedDisplay::Decimal);
+        let size_type = H2Number::new(SizedDefinition::U8, SizedDisplay::Decimal(Default::default()));
         let a: H2Type = LPString::new(size_type, UTF8::new())?;
         let array = a.resolve(offset, None)?;
 
@@ -181,7 +181,7 @@ mod tests {
 
     #[test]
     fn test_bad_type() -> SimpleResult<()> {
-        let size_type = H2Number::new(SizedDefinition::U8, SizedDisplay::Decimal);
+        let size_type = H2Number::new(SizedDefinition::U8, SizedDisplay::Decimal(Default::default()));
         assert!(LPString::new(size_type, IPv4::new(Endian::Big)).is_err());
 
         let size_type = IPv4::new(Endian::Big);
